Separate IVC generation from main ParaStell script
